refactor(pds): extract firehose framing and broadcast into repo/firehose
